Move Header component, update imports
